Update README.md
A banking app that allows the user to input their investment amounts and calculates how much interest they will gain over time.

Summarize the project and what problem it was solving.
The project was to make an interactive system that allowed the user to input their savings amount and rates and showed them how much interest they would gain over the amount of years they have inputted to see how their savings would grow. I had to create a system that allowed the user to go through a menu and input their own data into it for the results.

What did you do particularly well?
I was able to create a UI that allowed you to input whatever you wanted to create your banking results.

Where could you enhance your code? How would these improvements make your code more efficient, secure, and so on?
Once more comfortable with implementing other coding languages together with C++ it would improve my code a lot in terms of functionality and things I'm able to have the UI handle and the overall security of the program.

Which pieces of the code did you find most challenging to write, and how did you overcome this? What tools or resources are you adding to your support network?
Making sure I was setting up my for statements correctly to maintain the correct outputs. 

What skills from this project will be particularly transferable to other projects or course work?
The ability to work on troubleshooting which has been a big learning opportunity on this project and all of my others. And the ability to create interactive systems through C++.

How did you make this program maintainable, readable, and adaptable?
By properly spacing code and making it easy to read and comprehend the different functions, as well as maintaining formatting as a way to help you follow . The program is also set up to be able to be easily changed if needed to add onto it later on.